refactor: increase test coverage for mac for tty tests
